[sub-cookbooks] Move test recipes to separate cookbook.
