+++ Berksfile
@@ -18,3 +18,4 @@ cookbook "yum", path: "./third-party/yum-6.1.1"
 cookbook "yum-epel", path: "./third-party/yum-epel-4.1.2"
 cookbook "aws-parallelcluster-install", path: "./cookbooks/aws-parallelcluster-install"
 cookbook "aws-parallelcluster-config", path: "./cookbooks/aws-parallelcluster-config"
+cookbook "aws-parallelcluster-test", path: "./cookbooks/aws-parallelcluster-test"
