+++ cookbooks/aws-parallelcluster-config/metadata.rb
@@ -4,8 +4,8 @@ name 'aws-parallelcluster-config'
 maintainer 'Amazon Web Services'
 maintainer_email ''
 license 'Apache-2.0'
-description 'Installs/Configures AWS ParallelCluster'
-long_description 'Installs/Configures AWS ParallelCluster'
+description 'Configures AWS ParallelCluster'
+long_description 'Configures AWS ParallelCluster'
 issues_url 'https://github.com/aws/aws-parallelcluster-cookbook/issues'
 source_url 'https://github.com/aws/aws-parallelcluster-cookbook'
 chef_version '17.2.29'
