+++ cookbooks/aws-parallelcluster-test/metadata.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+name 'aws-parallelcluster-test'
+maintainer 'Amazon Web Services'
+maintainer_email ''
+license 'Apache-2.0'
+description 'Tests AWS ParallelCluster'
+long_description 'Tests AWS ParallelCluster'
+issues_url 'https://github.com/aws/aws-parallelcluster-cookbook/issues'
+source_url 'https://github.com/aws/aws-parallelcluster-cookbook'
+chef_version '17.2.29'
+version '3.0.0'
+
+supports 'amazon', '>= 2'
+supports 'centos', '>= 7'
+supports 'ubuntu', '>= 18.04'
