+++ cookbooks/aws-parallelcluster-test/recipes/test_envars.rb
