+++ cookbooks/aws-parallelcluster-test/recipes/test_imds.rb
