+++ cookbooks/aws-parallelcluster-test/recipes/test_openssh.rb
