+++ cookbooks/aws-parallelcluster-test/recipes/test_processes.rb
