+++ cookbooks/aws-parallelcluster-test/recipes/test_sudoers.rb
