+++ cookbooks/aws-parallelcluster-test/recipes/test_users.rb
